pyproject.toml
[tool.poetry]
name = "breathing_simulator"
version = "0.1.0"
description = ""
authors = ["chris <christopher.sutton12@gmail.com>"]

packages = [
	{ include="hardware", from="." },
	{ include="tests", from="." },
]

[tool.poetry.scripts]
blower_test = "tests.blower_tests:start"

[tool.poetry.dependencies]
python = "^3.7"
"RPi.GPIO" = "^0.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
